// src/app/pages/expenses/components/expenses-table/expenses.interface.ts
export interface ExpensesInterface {
    nickname: string;
    usuarioauth: string;
    claveauth: string;
    idempresa: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    costo: number;
    idstatususuario: number;
    emailsms: string;
    bfechainicial: boolean;
    fechainicial: string;
}